pipeline/requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langchain-pinecone>=0.2.0
langchain-core>=0.3.0
pinecone-client>=5.0.0
openai>=1.0.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
cohere>=5.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
flask>=3.0.0
